Moved code from tests/cpp/GMatrix.cpp to main/cpp/GMatrix.cpp. Declared GInterval class.
